services/backend/app/dtr/extractors/algorithmic.py
"""
Algorithmic Computer Vision Utilities

Non-LLM image processing for verification and enhancement.
Optional verification layer that can run alongside LLM analysis.

NOTE: These are optional enhancements. Pass 1 can work without them.
Implement when needed for higher precision.
"""
from typing import List, Tuple, Optional
import numpy as np
from PIL import Image
import io


def k_means_color_clustering(
    image_bytes: bytes,
    n_colors: int = 10
) -> List[Tuple[int, int, int]]:
    """
    Extract dominant colors using K-means clustering
    
    Args:
        image_bytes: Image data as bytes
        n_colors: Number of dominant colors to extract
    
    Returns:
        List of RGB tuples representing dominant colors
    
    NOTE: Requires scikit-learn. Implement when Pass 2 (surface) is needed.
    """
    # TODO: Implement when Pass 2 is built
    # from sklearn.cluster import KMeans
    # 
    # # Load image
    # image = Image.open(io.BytesIO(image_bytes))
    # image = image.convert('RGB')
    # 
    # # Reshape to array of pixels
    # pixels = np.array(image).reshape(-1, 3)
    # 
    # # K-means clustering
    # kmeans = KMeans(n_clusters=n_colors, random_state=42)
    # kmeans.fit(pixels)
    # 
    # # Get cluster centers (dominant colors)
    # colors = kmeans.cluster_centers_.astype(int)
    # 
    # return [tuple(c) for c in colors]
    
    raise NotImplementedError("K-means clustering not yet implemented")


def edge_detection_spacing(
    image_bytes: bytes
) -> List[int]:
    """
    Detect spacing values using edge detection
    
    Uses Canny edge detection to find element boundaries,
    then measures pixel distances between edges.
    
    Args:
        image_bytes: Image data as bytes
    
    Returns:
        List of detected spacing values in pixels
    
    NOTE: Requires opencv-python. Implement when needed for precision.
    """
    # TODO: Implement when higher precision spacing is needed
    # import cv2
    # 
    # # Load image
    # image = Image.open(io.BytesIO(image_bytes))
    # image_array = np.array(image.convert('L'))  # Grayscale
    # 
    # # Canny edge detection
    # edges = cv2.Canny(image_array, 100, 200)
    # 
    # # Detect spacing between edges
    # # ... (complex logic to measure gaps between edges)
    # 
    # return spacing_values
    
    raise NotImplementedError("Edge detection spacing not yet implemented")


def fft_grid_detection(
    image_bytes: bytes
) -> Optional[int]:
    """
    Detect underlying grid rhythm using FFT
    
    Uses Fast Fourier Transform to detect repeating spatial frequencies,
    revealing the base grid unit (4px, 8px, etc.).
    
    Args:
        image_bytes: Image data as bytes
    
    Returns:
        Detected grid quantum in pixels, or None if not detected
    
    NOTE: Advanced feature. Implement only if basic methods insufficient.
    """
    # TODO: Implement if needed for advanced grid detection
    # 
    # # Load image
    # image = Image.open(io.BytesIO(image_bytes))
    # image_array = np.array(image.convert('L'))
    # 
    # # Apply FFT
    # fft = np.fft.fft2(image_array)
    # fft_shifted = np.fft.fftshift(fft)
    # magnitude = np.abs(fft_shifted)
    # 
    # # Detect dominant frequencies
    # # ... (complex frequency analysis)
    # 
    # return detected_quantum
    
    raise NotImplementedError("FFT grid detection not yet implemented")


# ============================================================================
# PUBLIC API
# ============================================================================

__all__ = [
    'k_means_color_clustering',
    'edge_detection_spacing',
    'fft_grid_detection'
]
