services/backend/requirements.txt
numpy==2.0.2

# Core LLM providers
anthropic==0.75.0
google-generativeai>=0.8.3

# AWS
boto3==1.42.4

# Web framework
fastapi==0.123.5
pydantic==2.12.5
pydantic-settings==2.10.1
email-validator==2.2.0
mangum==0.18.0
websockets==12.0

# Authentication
PyJWT==2.10.1
python-jose==3.5.0
cryptography==46.0.3

# HTTP & utilities
requests==2.32.5
httpx==0.27.2
python-multipart==0.0.20
uvicorn==0.38.0
python-dotenv==1.2.1

# Image processing (for wireframe mode)
Pillow==10.4.0

# Playwright - DISABLED: Too large for Lambda zip, causes 250MB limit issues
# playwright==1.49.1
# playwright-stealth==1.0.6

# LLM Infrastructure (added for multi-provider support)

openai>=1.50.0  # Optional: For OpenAI GPT models (o-series reasoning)
aiofiles>=24.1.0  # Optional: Better async file handling